+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added new function `_add_file_name!` that makes sure that the file path/name of the distribution system being parsed in is added to the "files" dictionary of "pmd".
 - Fixed issues identified by unit tests related to new naming formatting for JSON files.
 - Added new JSON files with specific distribution system names to `test/json` folder.
+- Added `autorename.jl` unit tests designed to test errors that must appear when the new naming convention is not followed and `auto_rename=false`.
 
 ## v0.6.1